+++ src/lib.rs
@@ -6,3 +6,4 @@ pub mod path;
 pub mod rwlock;
 pub mod sqlite;
 pub mod tauri;
+pub mod time;